changed id type from int to Long added diffrent status for endpoints

+++ src/main/java/com/javPOL/magazineJava/dao/InvoiceDAO/InvoiceDao.java
@@ -1,8 +1,7 @@
 package com.javPOL.magazineJava.dao.InvoiceDAO;
 
 import com.javPOL.magazineJava.dao.Dao;
-import com.javPOL.magazineJava.model.Category;
 import com.javPOL.magazineJava.model.Invoice;
 
-public interface InvoiceDao extends Dao<Invoice, Integer> {
+public interface InvoiceDao extends Dao<Invoice, Long> {
 }
