+++ src/main/java/com/javPOL/magazineJava/dao/WarehouseProductDAO/WarehouseProductDao.java
@@ -3,5 +3,5 @@ package com.javPOL.magazineJava.dao.WarehouseProductDAO;
 import com.javPOL.magazineJava.dao.Dao;
 import com.javPOL.magazineJava.model.WarehouseProduct;
 
-public interface WarehouseProductDao extends Dao<WarehouseProduct, Integer> {
+public interface WarehouseProductDao extends Dao<WarehouseProduct, Long> {
 }
